+++ modules/apps/commerce/commerce-subscription-web/src/main/java/com/liferay/commerce/subscription/web/internal/constants/CommerceSubscriptionEntryScreenNavigationConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.subscription.web.internal.servlet.taglib.ui.constants;
+package com.liferay.commerce.subscription.web.internal.constants;
 
 /**
  * @author Alessio Antonio Rendina
